fix: Open file in edit mode intead of panicking
A file (text) can now be opened instead of panicking with an
'Not a directory' error.
Editing the file doesn't chnage the actual content as that is not
implemented yet.
If there is an error when opening a file, an error page is displayed
instead.

----
The error page only displays the error with no helpful message in it.

----
Drop use of Url and use PathBuf instead.
Using url makes filepaths to contain strange characters instead of
readable paths.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>