chore: Start dividing the view from the data
Instead of tightly coupling the view and the data, the two should now be
seperated.
This is to help in adding more data sources in the future without having
to make heavy UI chanages.

Remove the 'window/' directory as it was useless division.
Move it to the appropriate 'ui/' directory where its relevant.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>