chore: Start rewrite with `Hello World`
Rewrite the code with better organization with modularity in mind.
The old model made many assumptions that would prevent the application
from scaling.

The rewrite makes sure that the from and the backend are loosely coupled
with one not having to concern itself with he other.

---
The file system rewrite has began.
Now listing of the files and reading file content is supported.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>